pyproject.toml
[tool.poetry]
name = "NHNESymbolAutocomplete"
version = "0.1.0"
description = "Symbol autocomplete backend for nhne"
authors = ["Prasanta Kakati <prasantakakati1994@gmail.com>"]
license = "Not open source"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.4"
ipython = "^8.23.0"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
finnhub-python = "^2.4.20"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.4.0"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
